requirements.txt

viam-sdk==0.56.0

typing-extensions
